fix: jsx2mp template ast analyze
